Add files via upload
Oracle Hospitality Simphony on Prem database has two schemas, one is transaction schema and another one is security schema, when you are upgrading from an earlier version of Simphony to newer version of simphony there are some restrictions, Ex: If earlier version used microsdb schemas you cannot use same schemas to newer version, you would require remapping the schema before restoring the database. Importing tables in a different schema is very similar to importing tables in the same schema the only difference is that we have to resolve the ownership and tell the database about the new user of the table. In order to do so we use REMAP_SCHEMA parameter.